requirements.txt
amqp
anyio
appnope
asgiref
asttokens
backcall
billiard
cachetools
celery
Django
django-celery-beat
django-cors-headers
djangorestframework
djangorestframework-simplejwt
pika
Pillow
pytz
sqlparse
Werkzeug
whitenoise
psycopg2-binary
psycopg2
